test: move TestConvertStateGolden to only linux + mac
Windows runners are flaky for golden files
